Support /usr merge by allowing /bin/perl to be accelerated JB#36606
Now that /bin is a symlink to /usr/bin it seems that sometimes perl is
found as /bin/perl. This can manifest in things like autoreconf having
a shebang of #!/bin/perl (Speculatively because of PATH having /bin
first?)

This misses the /usr/bin/perl rules in a way which means that the ARM
perl tries to load x86 .so extensions.

Signed-off-by: David Greaves <david.greaves@jolla.com>